requirements.txt
altair==4.1.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
boto3==1.22.4
botocore==1.25.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
commonmark==0.9.1
contextvars==2.4
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
enum-compat==0.0.3
faiss-cpu==1.6.1
filelock==3.4.1
folium==0.2.1
googletrans==4.0.0rc1
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
nltk==3.6.7
notebook==6.4.10
numpy==1.19.2
packaging==21.3
pandas==1.1.2
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
ptyprocess==0.7.0
pycparser==2.21
pydeck==0.6.2
Pygments==2.12.0
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2022.4.24
requests==2.27.1
rfc3986==1.5.0
rich==12.4.1
s3transfer==0.5.2
sacremoses==0.0.49
scikit-learn==0.24.2
scipy==1.5.4
Send2Trash==1.8.0
sentence-transformers==0.3.8
sentencepiece==0.1.96
six==1.16.0
sniffio==1.2.0
streamlit==0.62.0
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tokenizers==0.8.1rc2
toml==0.10.2
toolz==0.11.2
torch==1.8.1
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
transformers==3.3.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.6.0
